+++ Tests/ksqlDB.RestApi.Client.Tests/Infrastructure/Extensions/TypeExtensionsTests.cs
@@ -1,7 +1,10 @@
 using System.Collections;
+using System.Linq.Expressions;
+using System.Reflection;
 using System.Text;
 using System.Text.Json.Serialization;
 using FluentAssertions;
+using ksqlDb.RestApi.Client.FluentAPI.Builders;
 using ksqlDB.RestApi.Client.Infrastructure.Extensions;
 using ksqlDB.RestApi.Client.KSql.Linq;
 using ksqlDB.RestApi.Client.KSql.Query;
@@ -357,6 +360,8 @@ public class TypeExtensionsTests
     attribute.Should().BeOfType<KeyAttribute>();
   }
 
+  #region GetMemberName
+
   private record MySensor
   {
     [JsonPropertyName("SensorId")]
@@ -373,7 +378,7 @@ public class TypeExtensionsTests
     var member = type.GetProperty(nameof(MySensor.Title));
 
     //Act
-    var memberName = member!.GetMemberName();
+    var memberName = member!.GetMemberName(new ModelBuilder());
 
     //Assert
     memberName.Should().Be(nameof(MySensor.Title));
@@ -388,9 +393,91 @@ public class TypeExtensionsTests
     //Act
     var member = type.GetProperty(nameof(MySensor.SensorId2));
 
-    var memberName = member!.GetMemberName();
+    var memberName = member!.GetMemberName(new ModelBuilder());
 
     //Assert
     memberName.Should().Be("SensorId");
   }
+
+  [Test]
+  public void GetMemberName_ModelBuilderHasColumnName()
+  {
+    //Arrange
+    var columnName = "Id";
+
+    var modelBuilder = new ModelBuilder();
+    modelBuilder.Entity<MySensor>()
+      .Property(c => c.SensorId2)
+      .HasColumnName(columnName);
+
+    var type = typeof(MySensor);
+
+    //Act
+    var member = type.GetProperty(nameof(MySensor.SensorId2));
+
+    var memberName = member!.GetMemberName(modelBuilder);
+
+    //Assert
+    memberName.Should().Be(columnName);
+  }
+
+  [Test]
+  public void GetMemberName_MemberExpression_ModelBuilderHasColumnName()
+  {
+    //Arrange
+    var columnName = "Id";
+    Expression<Func<MySensor, string>> expression = c => c.SensorId2;
+
+    var modelBuilder = new ModelBuilder();
+    modelBuilder.Entity<MySensor>()
+      .Property(expression)
+      .HasColumnName(columnName);
+
+    //Act
+    var memberName = ((MemberExpression)expression.Body).GetMemberName(modelBuilder);
+
+    //Assert
+    memberName.Should().Be(columnName);
+  }
+
+  [Test]
+  public void GetMemberName_FromMemberExpression_JsonPropertyNameAttributeWasUsed()
+  {
+    //Arrange
+    Expression<Func<MySensor, string>> expression = c => c.SensorId2;
+
+    var modelBuilder = new ModelBuilder();
+    modelBuilder.Entity<MySensor>()
+      .Property(expression)
+      .WithHeaders();
+    var memberExpression = (MemberExpression) expression.Body;
+
+    //Act
+    var memberName = memberExpression.GetMemberName(modelBuilder);
+
+    //Assert
+    var jsonPropertyNameAttribute = memberExpression.Member.GetCustomAttribute<JsonPropertyNameAttribute>();
+    memberName.Should().Be(jsonPropertyNameAttribute?.Name);
+  }
+
+  [Test]
+  public void GetMemberName_FromMemberExpression_PropertyNameWasUsed()
+  {
+    //Arrange
+    Expression<Func<MySensor, string>> expression = c => c.Title;
+
+    var modelBuilder = new ModelBuilder();
+    modelBuilder.Entity<MySensor>()
+      .Property(expression)
+      .WithHeaders();
+    var memberExpression = (MemberExpression)expression.Body;
+
+    //Act
+    var memberName = memberExpression.GetMemberName(modelBuilder);
+
+    //Assert
+    memberName.Should().Be(nameof(MySensor.Title));
+  }
+
+  #endregion
 }