+++ ksqlDb.RestApi.Client/KSql/RestApi/Extensions/MemberInfoExtensions.cs
@@ -1,4 +1,5 @@
 using System.Reflection;
+using ksqlDb.RestApi.Client.FluentAPI.Builders;
 using ksqlDB.RestApi.Client.KSql.RestApi.Enums;
 using ksqlDb.RestApi.Client.KSql.RestApi.Parsers;
 
@@ -11,7 +12,8 @@ namespace ksqlDB.RestApi.Client.KSql.RestApi.Extensions
     /// </summary>
     /// <param name="memberInfo"></param>
     /// <param name="escaping"></param>
+    /// <param name="modelBuilder"></param>
     /// <returns>the <c>memberInfo.Name</c> modified based on the provided <c>format</c></returns>
-    public static string Format(this MemberInfo memberInfo, IdentifierEscaping escaping) => IdentifierUtil.Format(memberInfo, escaping);
+    public static string Format(this MemberInfo memberInfo, IdentifierEscaping escaping, ModelBuilder modelBuilder) => IdentifierUtil.Format(memberInfo, escaping, modelBuilder);
   }
 }