+++ ksqlDb.RestApi.Client/KSql/RestApi/KSqlDbQueryStreamProvider.cs
@@ -1,6 +1,7 @@
 #if !NETSTANDARD
 using System.Net;
 using System.Text.Json;
+using ksqlDb.RestApi.Client.FluentAPI.Builders;
 using ksqlDb.RestApi.Client.KSql.Query.Context.Options;
 using ksqlDB.RestApi.Client.KSql.RestApi.Exceptions;
 using ksqlDB.RestApi.Client.KSql.RestApi.Responses;
@@ -12,8 +13,8 @@ namespace ksqlDB.RestApi.Client.KSql.RestApi
 {
   internal class KSqlDbQueryStreamProvider : KSqlDbProvider
   {
-    public KSqlDbQueryStreamProvider(IHttpClientFactory httpClientFactory, KSqlDbProviderOptions options, ILogger logger = null)
-      : base(httpClientFactory, options, logger)
+    public KSqlDbQueryStreamProvider(IHttpClientFactory httpClientFactory, IMetadataProvider metadataProvider, KSqlDbProviderOptions options, ILogger logger = null)
+      : base(httpClientFactory, metadataProvider, options, logger)
     {
 #if NETCOREAPP3_1
       AppContext.SetSwitch("System.Net.Http.SocketsHttpHandler.Http2UnencryptedSupport", true);